Share http client
